add inbox swipe actions
