+++ FlowCrypt/Controllers/Inbox/InboxViewController.swift
@@ -8,29 +8,31 @@ import FlowCryptUI
 
 @MainActor
 class InboxViewController: ViewController {
-    private lazy var logger = Logger.nested(Self.self)
+    lazy var logger = Logger.nested(Self.self)
 
     private let numberOfInboxItemsToLoad: Int
 
     let appContext: AppContextWithUser
     let tableNode: ASTableNode
 
-    private let decorator: InboxViewDecorator
-    private let messageOperationsApiClient: MessageOperationsApiClient
+    let decorator: InboxViewDecorator
+    let messageOperationsApiClient: MessageOperationsApiClient
+    let threadOperationsApiClient: MessagesThreadOperationsApiClient
+    let messageActionsHelper: MessageActionsHelper
     private let refreshControl = UIRefreshControl()
     private lazy var composeButton = ComposeButtonNode { [weak self] in
         self?.btnComposeTap()
     }
 
-    private let inboxDataApiClient: InboxDataApiClient
-    private let viewModel: InboxViewModel
-    private var inboxInput: [InboxItem] = []
+    let inboxDataApiClient: InboxDataApiClient
+    let viewModel: InboxViewModel
+    var inboxInput: [InboxItem] = []
     var state: InboxViewController.State = .idle
-    private var inboxTitle: String {
+    var inboxTitle: String {
         viewModel.folderName.isEmpty ? "Inbox" : viewModel.folderName
     }
 
-    private var shouldShowEmptyView: Bool {
+    var shouldShowEmptyView: Bool {
         inboxInput.isNotEmpty && (["SPAM", "TRASH"].contains(viewModel.path))
     }
 
@@ -59,6 +61,8 @@ class InboxViewController: ViewController {
 
         let mailProvider = try appContext.getRequiredMailProvider()
         self.messageOperationsApiClient = try mailProvider.messageOperationsApiClient
+        self.threadOperationsApiClient = try mailProvider.threadOperationsApiClient
+        self.messageActionsHelper = MessageActionsHelper(threadOperationsApiClient: threadOperationsApiClient)
         self.decorator = decorator
         self.tableNode = TableNode()
         self.isSearch = isSearch
@@ -418,323 +422,5 @@ extension InboxViewController {
 extension InboxViewController: Refreshable {
     func startRefreshing() {
         refresh()
-    }
-}
-
-extension InboxViewController: ASTableDataSource, ASTableDelegate {
-    func tableNode(_: ASTableNode, numberOfRowsInSection _: Int) -> Int {
-        switch state {
-        case .empty, .idle, .searchStart, .searching, .searchEmpty, .error:
-            return 1
-        case .fetched, .refresh, .fetching:
-            if shouldShowEmptyView, !inboxInput.isEmpty {
-                return inboxInput.count + 1
-            }
-            return inboxInput.count
-        }
-    }
-
-    func tableNode(_ tableNode: ASTableNode, nodeBlockForRowAt indexPath: IndexPath) -> ASCellNodeBlock {
-        cellNode(for: indexPath, and: visibleSize(for: tableNode))
-    }
-
-    func tableNode(_ tableNode: ASTableNode, didSelectRowAt indexPath: IndexPath) {
-        let inboxItemIndex = shouldShowEmptyView ? indexPath.row - 1 : indexPath.row
-
-        guard let inboxItem = inboxInput[safe: inboxItemIndex] else {
-            return
-        }
-
-        tableNode.deselectRow(at: indexPath, animated: true)
-
-        open(inboxItem: inboxItem, path: viewModel.path)
-    }
-
-    private func cellNode(for indexPath: IndexPath, and size: CGSize) -> ASCellNodeBlock {
-        return { [weak self] in
-            guard let self else { return ASCellNode() }
-
-            switch self.state {
-            case .empty:
-                return EmptyCellNode(
-                    input: self.decorator.emptyStateNodeInput(
-                        for: size,
-                        title: self.inboxTitle,
-                        imageName: self.viewModel.path.mailFolderIcon
-                    )
-                )
-            case .searchStart:
-                return TextCellNode(input: self.decorator.initialNodeInput(for: size, withSpinner: false))
-            case .searchEmpty:
-                return TextCellNode(input: self.decorator.searchEmptyStateNodeInput(for: size))
-            case .searching:
-                return TextCellNode.loading
-            case .idle:
-                let node = TextCellNode(input: self.decorator.initialNodeInput(for: size))
-                node.accessibilityIdentifier = "aid-inbox-idle-node"
-                return node
-            case .fetched, .refresh:
-                var rowNumber = indexPath.row
-                if self.shouldShowEmptyView {
-                    if indexPath.row == 0 {
-                        return self.emptyFolderNode()
-                    }
-                    rowNumber -= 1
-                }
-                guard let input = self.inboxInput[safe: rowNumber] else {
-                    return TextCellNode.loading
-                }
-                return InboxCellNode(input: .init(input))
-                    .then { $0.backgroundColor = .backgroundColor }
-            case .fetching:
-                guard let input = self.inboxInput[safe: indexPath.row] else {
-                    return TextCellNode.loading
-                }
-                return InboxCellNode(input: .init(input))
-            case let .error(message):
-                return TextCellNode(
-                    input: .init(
-                        backgroundColor: .backgroundColor,
-                        title: message,
-                        withSpinner: false,
-                        size: size
-                    )
-                )
-            }
-        }
-    }
-
-    private func emptyFolderNode() -> ASCellNode {
-        return EmptyFolderCellNode(
-            path: viewModel.path,
-            emptyFolder: { [weak self] in
-                self?.showConfirmAlert(
-                    message: "folder_empty_confirm".localized,
-                    onConfirm: { [weak self] _ in
-                        self?.emptyInboxFolder()
-                    }
-                )
-            }
-        )
-    }
-
-    private func emptyInboxFolder() {
-        Task {
-            do {
-                showSpinner()
-                try await self.messageOperationsApiClient.emptyFolder(path: viewModel.path)
-                self.state = .empty
-                self.inboxInput = []
-                await tableNode.reloadData()
-                hideSpinner()
-            } catch {
-                showAlert(message: error.errorMessage)
-            }
-        }
-    }
-}
-
-extension InboxViewController {
-    func getUpdatedIndex(for inboxItem: InboxItem) -> Int? {
-        let index = inboxInput.firstIndex(where: {
-            $0.title == inboxItem.title && $0.subtitle == inboxItem.subtitle && $0.type == inboxItem.type
-        })
-        logger.logInfo("Try to update inbox item at \(String(describing: index))")
-        return index
-    }
-
-    func updateMessage(isRead: Bool, at index: Int) {
-        guard inboxInput.count > index else { return }
-
-        logger.logInfo("Mark as read \(isRead) at \(index)")
-
-        // Mark wrapped message/thread(all mails in thread) as read/unread
-        inboxInput[index].markAsRead(isRead)
-        reloadMessage(index: index)
-    }
-
-    func updateMessage(labelsToAdd: [MessageLabel], labelsToRemove: [MessageLabel], at index: Int) {
-        guard inboxInput.count > index else { return }
-
-        inboxInput[index].update(labelsToAdd: labelsToAdd, labelsToRemove: labelsToRemove)
-        reloadMessage(index: index)
-    }
-
-    func reloadMessage(index: Int, animationDuration: Double = 0.3) {
-        DispatchQueue.main.asyncAfter(deadline: .now() + animationDuration) { [weak self] in
-            guard let self else { return }
-            self.tableNode.reloadRows(
-                at: [self.indexPathForMessage(at: index)],
-                with: .automatic
-            )
-        }
-    }
-
-    func removeMessage(at index: Int) {
-        guard inboxInput[safe: index] != nil else { return }
-
-        logger.logInfo("Try to remove at \(index)")
-        inboxInput.remove(at: index)
-
-        guard inboxInput.isNotEmpty else {
-            state = .empty
-            tableNode.reloadData()
-            return
-        }
-        switch state {
-        case let .fetched(.byNumber(total)):
-            let newTotalNumber = (total ?? 0) - 1
-            if newTotalNumber == 0 {
-                state = .empty
-                tableNode.reloadData()
-            } else {
-                state = .fetched(.byNumber(total: newTotalNumber))
-                do {
-                    try ObjcException.catch {
-                        self.tableNode.deleteRows(
-                            at: [self.indexPathForMessage(at: index)],
-                            with: .left
-                        )
-                    }
-                } catch {
-                    showAlert(message: "Failed to remove message at \(index) in fetched state: \(error)")
-                }
-            }
-        default:
-            do {
-                try ObjcException.catch {
-                    self.tableNode.deleteRows(
-                        at: [self.indexPathForMessage(at: index)],
-                        with: .left
-                    )
-                }
-            } catch {
-                showAlert(message: "Failed to remove message at \(index) in \(state): \(error)")
-            }
-        }
-    }
-
-    func open(inboxItem: InboxItem, path: String) {
-        if inboxItem.isDraft, let draft = inboxItem.messages.first {
-            open(draft: draft, appContext: appContext)
-        } else {
-            Task {
-                do {
-                    let viewController = try await ThreadDetailsViewController(
-                        appContext: appContext,
-                        inboxItem: inboxItem,
-                        onComposeMessageAction: { [weak self] action in
-                            guard let self else { return }
-
-                            switch action {
-                            case let .update(identifier), let .sent(identifier), let .delete(identifier):
-                                self.fetchUpdatedInboxItem(identifier: identifier)
-                            }
-                        },
-                        onComplete: { [weak self] action, message in
-                            self?.handleMessageOperation(message: message, action: action)
-                        }
-                    )
-                    navigationController?.pushViewController(viewController, animated: true)
-                } catch {
-                    showAlert(message: error.errorMessage)
-                }
-            }
-        }
-    }
-
-    private func open(draft: Message, appContext: AppContextWithUser) {
-        Task {
-            do {
-                let draftInfo = ComposeMessageInput.MessageQuoteInfo(
-                    message: draft,
-                    processed: nil
-                )
-
-                let controller = try await ComposeViewController(
-                    appContext: appContext,
-                    input: .init(type: .draft(draftInfo)),
-                    handleAction: { [weak self] action in
-                        switch action {
-                        case let .update(identifier):
-                            self?.fetchUpdatedInboxItem(identifier: identifier)
-                        case let .sent(identifier), let .delete(identifier):
-                            self?.deleteInboxItem(identifier: identifier)
-                        }
-                    }
-                )
-                navigationController?.pushViewController(controller, animated: true)
-            } catch {
-                showAlert(message: error.errorMessage)
-            }
-        }
-    }
-
-    private func fetchUpdatedInboxItem(identifier: MessageIdentifier) {
-        guard !inboxInput.isEmpty else {
-            fetchAndRenderEmails(nil)
-            return
-        }
-
-        Task {
-            guard let inboxItem = try await inboxDataApiClient.fetchInboxItem(
-                identifier: identifier,
-                path: path
-            ), !inboxItem.messages(with: path).isEmpty else {
-                deleteInboxItem(identifier: identifier)
-                return
-            }
-
-            guard let index = inboxInput.firstIndex(with: identifier) else {
-                inboxInput.insert(inboxItem, at: 0)
-                tableNode.insertRows(at: [indexPathForMessage(at: 0)], with: .automatic)
-                return
-            }
-
-            inboxInput[index] = inboxItem
-            tableNode.reloadRows(at: [indexPathForMessage(at: index)], with: .automatic)
-        }
-    }
-
-    private func deleteInboxItem(identifier: MessageIdentifier) {
-        guard let index = inboxInput.firstIndex(with: identifier) else { return }
-
-        inboxInput.remove(at: index)
-
-        if inboxInput.isEmpty {
-            state = .empty
-            tableNode.reloadData()
-        } else {
-            tableNode.deleteRows(at: [indexPathForMessage(at: index)], with: .automatic)
-        }
-    }
-
-    // MARK: Operation
-    private func indexPathForMessage(at index: Int) -> IndexPath {
-        let row = shouldShowEmptyView ? index + 1 : index
-        return IndexPath(row: row, section: 0)
-    }
-
-    private func handleMessageOperation(message: InboxItem, action: MessageAction) {
-        guard let indexToUpdate = getUpdatedIndex(for: message) else {
-            return
-        }
-
-        switch action {
-        case let .markAsRead(isRead):
-            updateMessage(isRead: isRead, at: indexToUpdate)
-        case .moveToTrash, .permanentlyDelete:
-            removeMessage(at: indexToUpdate)
-        case .archive, .moveToInbox:
-            if path.isEmpty { // no need to remove in 'All Mail' folder
-                updateMessage(
-                    labelsToAdd: action == .moveToInbox ? [.inbox] : [],
-                    labelsToRemove: action == .archive ? [.inbox] : [],
-                    at: indexToUpdate
-                )
-            } else {
-                removeMessage(at: indexToUpdate)
-            }
-        }
     }
 }
