+++ FlowCrypt/Controllers/Threads/Extensions/MessageActionsHandler.swift
@@ -43,8 +43,10 @@ extension MessageActionsHandler where Self: UIViewController {
     }
 
     private func setupNavigationBarItems(inboxItem: InboxItem, trashFolderPath: String?) {
-        logger.logInfo("setup navigation bar with \(trashFolderPath ?? "N/A")")
-        logger.logInfo("currentFolderPath \(currentFolderPath)")
+        logger.logInfo("""
+        setup navigation bar with \(trashFolderPath ?? "N/A")")
+        currentFolderPath \(currentFolderPath)
+        """)
 
         let helpButton = NavigationBarItemsView.Input(
             image: UIImage(systemName: "questionmark.circle"),
@@ -52,58 +54,44 @@ extension MessageActionsHandler where Self: UIViewController {
         ) { [weak self] in
             self?.handleInfoTap()
         }
-        let archiveButton = NavigationBarItemsView.Input(
-            image: UIImage(systemName: "tray.and.arrow.down"),
-            accessibilityId: "aid-archive-button"
-        ) { [weak self] in
-            self?.handleArchiveTap()
-        }
-        let moveToInboxButton = NavigationBarItemsView.Input(
-            image: UIImage(systemName: "tray.and.arrow.up"),
-            accessibilityId: "aid-move-to-inbox-button"
-        ) { [weak self] in
-            self?.handleMoveToInboxTap()
-        }
-        let trashButton = NavigationBarItemsView.Input(
-            image: UIImage(systemName: "trash"),
-            accessibilityId: "aid-delete-button"
-        ) { [weak self] in
-            self?.handleTrashTap()
-        }
-        let unreadButton = NavigationBarItemsView.Input(
-            image: UIImage(systemName: "envelope"),
-            accessibilityId: "aid-read-button"
-        ) { [weak self] in
-            self?.handleMarkUnreadTap()
-        }
 
-        var items: [NavigationBarItemsView.Input]
-
+        var actions: [MessageAction]
         switch currentFolderPath.lowercased() {
         case trashFolderPath?.lowercased():
-            logger.logInfo("trash - helpButton, moveToInboxButton, trashButton")
-            // in case we are in trash folder ([Gmail]/Trash or Deleted for Outlook, etc)
-            // we need to have only help, 'move to inbox' and trash buttons
-            items = [helpButton, moveToInboxButton, trashButton]
+            actions = [.moveToInbox, .moveToTrash]
         case "draft":
-            // for Gmail inbox we also need to have archive and unread buttons
-            logger.logInfo("draft - helpButton, trashButton")
-            items = [helpButton, trashButton]
+            actions = [.moveToTrash]
         default:
-            // in any other folders
-            items = [helpButton, trashButton, unreadButton]
+            actions = [.moveToTrash, .markAsUnread]
             if inboxItem.isInbox {
-                logger.logInfo("inbox - helpButton, archiveButton, trashButton, unreadButton")
-                items.insert(archiveButton, at: 1)
+                actions.insert(.archive, at: 0)
             } else if inboxItem.shouldShowMoveToInboxButton {
-                logger.logInfo("archive - helpButton, moveToInboxButton, trashButton, unreadButton")
-                items.insert(moveToInboxButton, at: 1)
-            } else {
-                logger.logInfo("sent - helpButton, trashButton, unreadButton")
+                actions.insert(.moveToInbox, at: 0)
             }
         }
 
+        let items = [helpButton] + actions.map { createNavigationBarButton(action: $0) }
         navigationItem.rightBarButtonItem = NavigationBarItemsView(with: items)
+    }
+
+    private func createNavigationBarButton(action: MessageAction) -> NavigationBarItemsView.Input {
+        .init(
+            image: action.image,
+            accessibilityId: action.accessibilityIdentifier
+        ) { [weak self] in
+            switch action {
+            case .moveToTrash, .permanentlyDelete:
+                self?.handleTrashTap()
+            case .moveToInbox:
+                self?.handleMoveToInboxTap()
+            case .archive:
+                self?.handleArchiveTap()
+            case .markAsUnread:
+                self?.handleMarkUnreadTap()
+            case .markAsRead:
+                break
+            }
+        }
     }
 
     func handleInfoTap() {
