+++ FlowCrypt/Controllers/Threads/Models/MessageAction.swift
@@ -6,20 +6,61 @@
 //  Copyright © 2017-present FlowCrypt a. s. All rights reserved.
 //
 
-import Foundation
+import UIKit
 
 typealias MessageActionCompletion = (MessageAction, InboxItem) -> Void
 
-enum MessageAction: Equatable {
-    case moveToTrash, moveToInbox, archive, markAsRead(Bool), permanentlyDelete
+enum MessageAction: String, Equatable {
+    case moveToTrash, moveToInbox, archive, markAsRead, markAsUnread, permanentlyDelete
 
-    var text: String? {
+    var imageName: String {
+        switch self {
+        case .moveToTrash, .permanentlyDelete:
+            return "trash"
+        case .moveToInbox:
+            return "tray.and.arrow.up"
+        case .archive:
+            return "tray.and.arrow.down"
+        case .markAsUnread:
+            return "envelope.open"
+        case .markAsRead:
+            return "envelope"
+        }
+    }
+
+    var image: UIImage? { UIImage(systemName: imageName) }
+
+    var color: UIColor {
+        switch self {
+        case .moveToTrash, .permanentlyDelete:
+            return .red
+        case .moveToInbox, .archive, .markAsRead, .markAsUnread:
+            return .main
+        }
+    }
+
+    var accessibilityIdentifier: String {
+        switch self {
+        case .moveToTrash, .permanentlyDelete:
+            return "aid-delete-button"
+        case .moveToInbox:
+            return "aid-move-to-inbox-button"
+        case .archive:
+            return "aid-archive-button"
+        case .markAsRead:
+            return "aid-read-button"
+        case .markAsUnread:
+            return "aid-unread-button"
+        }
+    }
+
+    var successMessage: String? {
         switch self {
         case .moveToTrash: return "email_removed".localized
         case .moveToInbox: return "email_moved_to_inbox".localized
         case .archive: return "email_archived".localized
         case .permanentlyDelete: return "email_deleted".localized
-        case .markAsRead: return nil
+        case .markAsRead, .markAsUnread: return nil
         }
     }
 
@@ -29,7 +70,7 @@ enum MessageAction: Equatable {
         case .moveToInbox: return "error_move_inbox".localized
         case .archive: return "error_archive".localized
         case .permanentlyDelete: return "error_permanently_delete".localized
-        case .markAsRead: return nil
+        case .markAsRead, .markAsUnread: return nil
         }
     }
 }
