--- FlowCrypt/Core/RSA/ClearMessage.swift
@@ -1,97 +1,0 @@
-//
-//  ClearMessage.swift
-//  SwiftyRSA
-//
-//  Created by Lois Di Qual on 5/18/17.
-//  Copyright © 2017 Scoop. All rights reserved.
-//
-
-import Foundation
-
-public class ClearMessage: RSAMessage {
-
-    /// Data of the message
-    public let data: Data
-
-    /// Creates a clear message with data.
-    ///
-    /// - Parameter data: Data of the clear message
-    public required init(data: Data) {
-        self.data = data
-    }
-
-    /// Creates a clear message from a string, with the specified encoding.
-    ///
-    /// - Parameters:
-    ///   - string: String value of the clear message
-    ///   - encoding: Encoding to use to generate the clear data
-    /// - Throws: SwiftyRSAError
-    public convenience init(string: String, using encoding: String.Encoding) throws {
-        guard let data = string.data(using: encoding) else {
-            throw SwiftyRSAError.stringToDataConversionFailed
-        }
-        self.init(data: data)
-    }
-
-    /// Returns the string representation of the clear message using the specified
-    /// string encoding.
-    ///
-    /// - Parameter encoding: Encoding to use during the string conversion
-    /// - Returns: String representation of the clear message
-    /// - Throws: SwiftyRSAError
-    public func string(encoding: String.Encoding) throws -> String {
-        guard let str = String(data: data, encoding: encoding) else {
-            throw SwiftyRSAError.dataToStringConversionFailed
-        }
-        return str
-    }
-
-    /// Encrypts a clear message with a public key and returns an encrypted message.
-    ///
-    /// - Parameters:
-    ///   - key: Public key to encrypt the clear message with
-    ///   - padding: Padding to use during the encryption
-    /// - Returns: Encrypted message
-    /// - Throws: SwiftyRSAError
-    public func encrypted(with key: PublicKey, padding: Padding) throws -> EncryptedMessage {
-
-        let blockSize = SecKeyGetBlockSize(key.reference)
-
-        var maxChunkSize: Int
-        switch padding {
-        case []:
-            maxChunkSize = blockSize
-        case .OAEP:
-            maxChunkSize = blockSize - 42
-        default:
-            maxChunkSize = blockSize - 11
-        }
-
-        var decryptedDataAsArray = [UInt8](repeating: 0, count: data.count)
-        (data as NSData).getBytes(&decryptedDataAsArray, length: data.count)
-
-        var encryptedDataBytes = [UInt8](repeating: 0, count: 0)
-        var idx = 0
-        while idx < decryptedDataAsArray.count {
-
-            let idxEnd = min(idx + maxChunkSize, decryptedDataAsArray.count)
-            let chunkData = [UInt8](decryptedDataAsArray[idx ..< idxEnd])
-
-            var encryptedDataBuffer = [UInt8](repeating: 0, count: blockSize)
-            var encryptedDataLength = blockSize
-
-            let status = SecKeyEncrypt(key.reference, padding, chunkData, chunkData.count, &encryptedDataBuffer, &encryptedDataLength)
-
-            guard status == noErr else {
-                throw SwiftyRSAError.chunkEncryptFailed(index: idx)
-            }
-
-            encryptedDataBytes += encryptedDataBuffer
-
-            idx += maxChunkSize
-        }
-
-        let encryptedData = Data(bytes: encryptedDataBytes, count: encryptedDataBytes.count)
-        return EncryptedMessage(data: encryptedData)
-    }
-}
